Refactor the workout history page, other ui tweaks... some for better, some for worse
